ObjectiveXMPP:
Using a queuedMessages property inside the XMPPConnection; I still need to get around to changing all of the ivar accesses I inherited from the xmppframework code. That must be one of the last pieces.
Added constants for the common stanzas.
Added a function to determine if a <message/> is being composed, so that it can be excluded from the interface etc.